requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
pandas-ta>=0.3.14b0
aiohttp>=3.9.0

# MetaAPI requirements
metaapi-cloud-sdk>=27.0.0

# Additional commonly needed packages for trading bots
ccxt>=4.0.0
asyncio-throttle>=1.0.0
aiosqlite>=0.19.0